--- a/tests/Jackalope/Transport/DoctrineDBAL/Query/QOMWalkerTest.php
+++ b/tests/Jackalope/Transport/DoctrineDBAL/Query/QOMWalkerTest.php
@@ -40,7 +40,7 @@ class QOMWalkerTest extends TestCase
         ;
         $this->factory = new QueryObjectModelFactory(new Factory);
         $this->walker = new QOMWalker($this->nodeTypeManager, $conn);
-        $this->defaultColumns = 'n0.id AS n0_id, n0.path AS n0_path, n0.parent AS n0_parent, n0.local_name AS n0_local_name, n0.namespace AS n0_namespace, n0.workspace_name AS n0_workspace_name, n0.identifier AS n0_identifier, n0.type AS n0_type, n0.props AS n0_props, n0.depth AS n0_depth, n0.sort_order AS n0_sort_order';
+        $this->defaultColumns = 'n0.id AS n0_id, n0.path AS n0_path, n0.parent AS n0_parent, n0.local_name AS n0_local_name, n0.namespace AS n0_namespace, n0.workspace_name AS n0_workspace_name, n0.identifier AS n0_identifier, n0.type AS n0_type, n0.props AS n0_props, n0.numerical_props AS n0_numerical_props, n0.depth AS n0_depth, n0.sort_order AS n0_sort_order';
     }
 
     public function testDefaultQuery()
@@ -186,7 +186,7 @@ class QOMWalkerTest extends TestCase
         $this->assertEquals(sprintf("SELECT %s FROM phpcr_nodes n0 WHERE n0.workspace_name = ? AND n0.type IN ('nt:unstructured') AND n0.path $op '/'", $this->defaultColumns), $sql);
     }
 
-    public function testQueryWithOrderings()
+    public function testQueryWithPathOrder()
     {
         $this->nodeTypeManager->expects($this->once())->method('getSubtypes')->will($this->returnValue( array() ));
 
@@ -202,6 +202,49 @@ class QOMWalkerTest extends TestCase
         $this->assertEquals(
             sprintf("SELECT %s FROM phpcr_nodes n0 WHERE n0.workspace_name = ? AND n0.type IN ('nt:unstructured') ORDER BY n0.path ASC", $this->defaultColumns),
             $sql
+        );
+    }
+
+    public function testQueryWithOrderings()
+    {
+        $driver = $this->conn->getDriver()->getName();
+
+        $this->nodeTypeManager->expects($this->once())->method('getSubtypes')->will($this->returnValue( array() ));
+
+        $query = $this->factory->createQuery(
+            $this->factory->selector('nt:unstructured', 'nt:unstructured'),
+            null,
+            array($this->factory->ascending($this->factory->propertyValue('nt:unstructured', "foobar"))),
+            array()
+        );
+
+        $res = $this->walker->walkQOMQuery($query);
+
+
+        switch ($driver) {
+            case 'pdo_pgsql':
+                $ordering = 
+                    "CAST((xpath('//sv:property[@sv:name=\"foobar\"]/sv:value[1]/text()', CAST(n0.numerical_props AS xml), ARRAY[ARRAY['sv', 'http://www.jcp.org/jcr/sv/1.0']]))[1]::text AS DECIMAL), " . 
+                   "(xpath('//sv:property[@sv:name=\"foobar\"]/sv:value[1]/text()', CAST(n0.props AS xml), ARRAY[ARRAY['sv', 'http://www.jcp.org/jcr/sv/1.0']]))[1]::text ASC";
+                break;
+            default:
+                $ordering = 
+                    "CAST(EXTRACTVALUE(n0.numerical_props, '//sv:property[@sv:name=\"foobar\"]/sv:value[1]') AS DECIMAL), " .
+                    "EXTRACTVALUE(n0.props, '//sv:property[@sv:name=\"foobar\"]/sv:value[1]') ASC";
+        }
+
+
+        $this->assertEquals(
+            sprintf(
+                implode(' ', array(
+                    "SELECT %s FROM phpcr_nodes n0 WHERE n0.workspace_name = ?",
+                    "AND n0.type IN ('nt:unstructured')",
+                    "ORDER BY",
+                    $ordering
+                )),
+                $this->defaultColumns
+            ),
+            $res[2]
         );
     }