@dbu 2015-01-04
the travis fail is only the postgres problem. +1 to merge from me.

(and maybe we should skip the failing test when the db is postgres and the env travis, if we can't find the problem, to have a reliable build)